+++ src/UC_Quantum_Lab/_src.py
@@ -1,16 +1,5 @@
-from . import _config_dir, _trigger_file
+from . import _config_dir
 import os
-
-"""
-Creates a file that triggers the vscode extension
-
-"""
-def _trigger():
-    global _config_dir, _trigger_file
-    
-    if _config_dir in os.listdir():
-        with open(os.path.join(_config_dir, _trigger_file), 'w'): pass
-
 
 """
 prepends inputted path with the config directory if it exists