+++ src/UC_Quantum_Lab/layout.py
@@ -1,5 +1,4 @@
-from . import _states, _circs, _hists, _layout_file, _master_show, _config_dir, _trigger_file
-from ._src import _trigger
+from . import _states, _circs, _hists, _layout_file, _master_show
 import os, json, platform
 
 _layout = {}
@@ -17,6 +16,7 @@ def _image_list_to_str(image_list:list[str])->str:
         if platform.system() == "Windows":
             p = p[p.index(":")+1:].replace("\\", "/")
         to_return+=f"<img src=\"{{URI}}{p}\" alt=\"no image to display\">"
+        #to_return+=f"<div class=\"image-container\">\n<div class=\"image detail-view\" style=\"background-image:url({{URI}}{p});\"></div>\n</div>\n"
     return to_return
 
 def _inverter(layout, flip:dict[str,str]):
@@ -133,14 +133,4 @@ def _run():
 def _layout_at_exit():
     from . import _master_show
     if _master_show:
-        #print("here")
-        from .layout import _run
         _run()
-        _trigger()
-
-def _exit():
-    if _config_dir in os.listdir():
-        for item in os.listdir(_config_dir):
-            # deletes png html or the trigger file from the config dir
-            if item == _trigger_file:
-                os.remove(os.path.join(_config_dir, item))